% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_plot.R
\name{accuracy_plot}
\alias{accuracy_plot}
\title{Produces a plot for the accuracy}
\usage{
accuracy_plot(sample = NULL, population, method = "higher")
}
\arguments{
\item{sample}{sample}

\item{population}{population}

\item{method}{method. Default is 'higher'}
}
\description{
Produces a plot for the accuracy
}
