% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_utils.R
\name{artfish_accuracy}
\alias{artfish_accuracy}
\title{artfish_accuracy}
\usage{
artfish_accuracy(n, N, method = "higher")
}
\arguments{
\item{n}{n}

\item{N}{N}

\item{method}{method}
}
\description{
artfish_accuracy
}
