% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_report.R
\name{artfish_compute_report}
\alias{artfish_compute_report}
\title{artfish_compute_report}
\usage{
artfish_compute_report(
  effort,
  effort_source,
  active_vessels,
  active_vessels_strategy,
  active_days,
  landings,
  minor_strata
)
}
\arguments{
\item{effort}{effort}

\item{effort_source}{effort_source}

\item{active_vessels}{active_vessels}

\item{active_vessels_strategy}{active_vessels_strategy}

\item{active_days}{active_days}

\item{landings}{landings}

\item{minor_strata}{minor_strata}
}
\value{
the result of Artfish
}
\description{
artfish_compute_report
}
