% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_new.R
\name{artfish_new}
\alias{artfish_new}
\title{artfish_new}
\usage{
artfish_new(
  active_vessels = NULL,
  effort = NULL,
  effort_source = c("fisher_interview", "boat_counting"),
  active_days = NULL,
  landings = NULL,
  minor_strata = NULL
)
}
\arguments{
\item{active_vessels}{active vessels}

\item{effort}{effort}

\item{effort_source}{effort source whether it's derived from -B1- (fishers interviews) 
or -B2- (boat counting)}

\item{active_days}{active days}

\item{landings}{landings}

\item{minor_strata}{minor_strata (to investigate further later)}
}
\value{
the result of Artfish
}
\description{
artfish_new
}
