% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_new.R
\name{artfish_new_by_period}
\alias{artfish_new_by_period}
\title{artfish_new_by_period}
\usage{
artfish_new_by_period(
  year = NULL,
  month = NULL,
  active_vessels,
  effort,
  effort_source = c("fisher_interview", "boat_counting"),
  active_days = NULL,
  landings,
  minor_strata = NULL,
  validate = TRUE
)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{active_vessels}{active vessels}

\item{effort}{effort}

\item{effort_source}{effort source whether it's derived from -B1- (fishers interviews) 
or -B2- (boat counting)}

\item{active_days}{active days}

\item{landings}{landings}

\item{minor_strata}{minor_strata (to investigate further later)}

\item{validate}{validate}
}
\value{
the result of Artfish for a given year/month
}
\description{
artfish_new_by_period
}
