% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_accuracy.R
\name{artfish_shiny_accuracy_server}
\alias{artfish_shiny_accuracy_server}
\title{POC shiny module server}
\usage{
artfish_shiny_accuracy_server(id, lang = NULL)
}
\arguments{
\item{id}{id}

\item{lang}{lang a reactive version of the language. Default is \code{NULL} (optional)}
}
\description{
Accuracy toolbox shiny module server
}
