% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_accuracy.R
\name{artfish_shiny_accuracy_ui}
\alias{artfish_shiny_accuracy_ui}
\title{POC shiny module UI}
\usage{
artfish_shiny_accuracy_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
POC shiny module UI
}
