% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_computation.R
\name{artfish_shiny_computation_server}
\alias{artfish_shiny_computation_server}
\title{ARTFISH Computation - Shiny Server Module}
\usage{
artfish_shiny_computation_server(
  id,
  refresh,
  effort,
  effort_source,
  active_vessels,
  active_vessels_strategy,
  census_typology,
  active_days,
  landings,
  minor_strata = NULL,
  progress_fn = NULL
)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}

\item{refresh}{refresher for computation}

\item{effort}{effort data}

\item{effort_source}{Character string indicating the type of effort source.}

\item{active_vessels}{active vessels data}

\item{active_vessels_strategy}{active vessels strategy}

\item{census_typology}{census typology}

\item{active_days}{active days}

\item{landings}{landings}

\item{minor_strata}{Character string targeting a column name considered as minor strata.}

\item{progress_fn}{A function to monitor progress}
}
\description{
Server-side shiny logic to run Artfish computation as business service
}
