% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_fishing_unit.R
\name{artfish_shiny_fishing_unit_ui}
\alias{artfish_shiny_fishing_unit_ui}
\title{ARTFISH Fishing unit - Shiny UI Module}
\usage{
artfish_shiny_fishing_unit_ui(id)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}
}
\description{
User interface for the ARTFISH Fishing Unit Dashboard
}
