% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_overview.R
\name{artfish_shiny_overview_server}
\alias{artfish_shiny_overview_server}
\title{ARTFISH Overview - Shiny Server Module}
\usage{
artfish_shiny_overview_server(
  id,
  lang = NULL,
  estimate,
  effort_source = NULL,
  minor_strata = NULL,
  opts = list()
)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}

\item{lang}{Optional language parameter.
Can be either:
\itemize{
  \item a character string (static language), or
  \item a reactive returning a character string (dynamic language)
}
If \code{NULL}, the current global language is used.
Default is \code{NULL}}

\item{estimate}{A reactive data frame aggregating the output of \code{artfish_compute}, enriched with human-readable labels.}

\item{effort_source}{Reactive Character string indicating the type of effort source.
Must be either \code{"fisher_interview"} or \code{"boat_counting"}.
Not activated}

\item{minor_strata}{Reactive Character string targeting a column name considered as minor strata.
Not activated}

\item{opts}{a named list of options. 
For now only supports the \code{refresh_ui} that gives the capacity to inject a refresh UI button (for dynamic computation)}
}
\description{
Server-side logic for the ARTFISH Overview dashboard.

The *"Total catch and effort - Statistics at country level"* module provides a high-level visualization of key output metrics derived from \code{artfish_compute_report}. 
It acts as a dashboard to track the overall evolution of fishing activity over time, with the ability to filter by fishing units and time range.

The module is composed of:
\itemize{
  \item A selector panel to adjust the time range and fishing units to display
  \item A set of KPIs presenting aggregated country-level metrics for the selected period
  \item Seven boxed plots powered by \code{fdishinyr::generic_chart} allowing users to analyse:
  \itemize{
    \item (1) Total catch estimation
    \item (2) Species composition (by quantity)
    \item (3) Total nominal effort
    \item (4) Number of active vessels
    \item (5) Total value estimation
    \item (6) Species composition (by value)
    \item (7) Catch per unit effort (CPUE)
  }
}
}
