% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_overview.R
\name{artfish_shiny_overview_ui}
\alias{artfish_shiny_overview_ui}
\title{ARTFISH Overview - Shiny UI Module}
\usage{
artfish_shiny_overview_ui(id)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}
}
\description{
User interface for the ARTFISH Overview Dashboard
}
