% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_report.R
\name{artfish_shiny_report_ui}
\alias{artfish_shiny_report_ui}
\title{ARTFISH Report - Shiny UI Module}
\usage{
artfish_shiny_report_ui(id)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}
}
\description{
User interface for the ARTFISH Report Dashboard
}
