% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_species.R
\name{artfish_shiny_species_server}
\alias{artfish_shiny_species_server}
\title{ARTFISH Species - Shiny Server Module}
\usage{
artfish_shiny_species_server(
  id,
  lang = NULL,
  estimate,
  effort_source = NULL,
  minor_strata = NULL
)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}

\item{lang}{Optional language parameter.
Can be either:
\itemize{
  \item a character string (static language), or
  \item a reactive returning a character string (dynamic language)
}
If \code{NULL}, the current global language is used.
Default is \code{NULL}}

\item{estimate}{A data frame aggregating the output of \code{artfish_compute}, enriched with human-readable labels.}

\item{effort_source}{Character string indicating the type of effort source.
Must be either \code{"fisher_interview"} or \code{"boat_counting"}.
Not activated}

\item{minor_strata}{Character string targeting a column name considered as minor strata.
Not activated}
}
\description{
Server-side logic for the ARTFISH Species dashboard.

The *"Catch and effort - Detailed by species"* module provides a species-oriented workspace for exploring key output metrics derived from \code{artfish_compute_report}, with a focus on individual species caught.

The initial view displays a selector inviting the user to choose a species. 
The dropdown lists all species for which entries are available in the database. 
Users can then analyse the evolution of fishing activity over time, with filtering options for species, fishing units, and time range.

The module is composed of:
\itemize{
  \item A selector panel to choose the species to display
  \item A selector panel to adjust the time range and fishing units to display
  \item Two boxed plots powered by \code{fdishinyr::generic_chart} allowing users to analyse, at species level:
  \itemize{
    \item (1) Fishing unit composition (by quantity)
    \item (2) Target species ranking (by quantity)
  }
  \item A set of KPIs presenting aggregated metrics for the selected period
  \item Five boxed plots powered by \code{fdishinyr::generic_chart} allowing users to analyse key Artfish metrics:
  \itemize{
    \item (1) Total catch estimation
    \item (2) Catch per unit effort (CPUE)
    \item (3) Total value estimation
    \item (4) Average price
    \item (5) Total nominal effort
  }
}
}
