% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_species.R
\name{artfish_shiny_species_ui}
\alias{artfish_shiny_species_ui}
\title{ARTFISH Species - Shiny UI module}
\usage{
artfish_shiny_species_ui(id)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}
}
\description{
User interface for the ARTFISH Species Dashboard
}
