% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfishr.R
\docType{package}
\name{artfishr}
\alias{artfishr}
\alias{artfishr-package}
\title{Tools for implementing ArtFish methodology}
\description{
artfishr provides a set of utility tools to implement ArtFish methodology
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdiwg/artfishr}
  \item Report bugs at \url{https://github.com/fdiwg/artfishr/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
