% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny.R
\name{artfishr_shiny_server}
\alias{artfishr_shiny_server}
\alias{artfish_shiny_server}
\title{Artfishr shiny server function}
\usage{
artfish_shiny_server(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Artfishr shiny server function
}
