% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny.R
\name{artfishr_shiny_ui}
\alias{artfishr_shiny_ui}
\alias{artfish_shiny_ui}
\title{Artfishr shiny ui function}
\usage{
artfish_shiny_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Artfishr shiny ui function
}
