% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_active_days.R
\name{complete_active_days}
\alias{complete_active_days}
\title{Complete the active days.}
\usage{
complete_active_days(active_days)
}
\arguments{
\item{active_days}{active days object of class \link[tibble]{tibble}}
}
\value{
the curated active days table
}
\description{
If it is set to zero or or is NA, the nb of active days
in the period will be auto-generated
}
