% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cpue_per_trip_statistics.R
\name{compute_cpue_per_trip_statistics}
\alias{compute_cpue_per_trip_statistics}
\title{Compute CPUE per trip descriptive statistics}
\usage{
compute_cpue_per_trip_statistics(
  landings,
  minor_strata = NULL,
  confidence = 0.9,
  keep_values = TRUE
)
}
\arguments{
\item{landings}{A standardized landings dataset.}

\item{minor_strata}{Optional column name defining the minor strata. Default is \code{NULL}.}

\item{confidence}{Confidence level used to compute confidence intervals.Default is \code{0.90}.}

\item{keep_values}{Logical. Should the original values be returned in the output? Default is \code{TRUE}.}
}
\value{
A tibble containing descriptive statistics.
}
\description{
Computes descriptive and precision statistics for fishing-trip CPUE values, aggregated by year, month, fishing unit and optionally by minor strata.
This function relies on the generic descriptive statistics engine \code{compute_summary_statistics()}.

Unlike \code{compute_cpue()}, which returns aggregated CPUE estimates, this function summarises the distribution of individual fishing-trip CPUE observations.
}
