% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_new.R
\name{compute_effort_activity_coefficient}
\alias{compute_effort_activity_coefficient}
\title{Computes effort activity coefficient}
\usage{
compute_effort_activity_coefficient(effort)
}
\arguments{
\item{effort}{effort data}
}
\value{
the activity coefficient by strata
}
\description{
Computes effort activity coefficient
}
