% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_effort_activity_coefficient_per_observation_statistics.R
\name{compute_effort_activity_coefficient_per_observation_statistics}
\alias{compute_effort_activity_coefficient_per_observation_statistics}
\title{Compute activity coefficient descriptive statistics}
\usage{
compute_effort_activity_coefficient_per_observation_statistics(
  effort,
  effort_source = c("fisher_interview", "boat_counting"),
  minor_strata = NULL,
  confidence = 0.9,
  keep_values = TRUE
)
}
\arguments{
\item{effort}{A standardized effort dataset.}

\item{effort_source}{Source of effort observations.}

\item{minor_strata}{Optional column name defining the minor strata. Default is \code{NULL}.}

\item{confidence}{Confidence level used to compute confidence intervals.Default is \code{0.90}.}

\item{keep_values}{Logical. Should the original values be returned in the output? Default is \code{TRUE}.}
}
\value{
A tibble containing descriptive statistics.
}
\description{
Computes descriptive and precision statistics on individual activity coefficient values.
This function relies on the generic descriptive statistics engine \code{compute_summary_statistics()}.

Unlike \code{compute_effort_activity_coefficient()}, which returns aggregated activity coefficient estimates, this function summarises the distribution of coefficient values at observation level.

For fisher interview data, one observation corresponds to one fisher interview and the activity coefficient is computed as the ratio between fishing duration and the fishing reference period.

For boat counting data, one observation corresponds to one boat counting event and the activity coefficient is computed as the ratio between fleet engagement number and fleet engagement maximum.
}
