% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_effort_estimate.R
\name{compute_effort_estimate}
\alias{compute_effort_estimate}
\title{Computes nominal effort estimate}
\usage{
compute_effort_estimate(
  effort,
  effort_source = c("fisher_interview", "boat_counting", "household_interview"),
  landings,
  active_days = NULL,
  active_vessels = NULL,
  active_vessels_strategy = c("latest", "closest"),
  census_typology = NULL,
  minor_strata = NULL,
  progress_fn = NULL
)
}
\arguments{
\item{effort}{effort}

\item{effort_source}{effort_source}

\item{landings}{landings}

\item{active_days}{active_days}

\item{active_vessels}{active_vessels}

\item{active_vessels_strategy}{active_vessels_strategy}

\item{census_typology}{census_typology}

\item{minor_strata}{minor_strata}

\item{progress_fn}{progress_fn}
}
\description{
Computes nominal effort estimate
}
