% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/artfish_compute_effort_estimate_with_IRD_Pechart.R
\name{compute_effort_estimate_with_IRD_Pechart}
\alias{compute_effort_estimate_with_IRD_Pechart}
\title{Computes nominal effort estimate}
\usage{
compute_effort_estimate_with_IRD_Pechart(
  effort,
  effort_source = c("household_interview"),
  landings,
  active_days,
  census_typology,
  minor_strata = NULL,
  progress_fn = NULL
)
}
\arguments{
\item{effort}{effort}

\item{effort_source}{effort_source}

\item{landings}{landings}

\item{active_days}{active_days}

\item{census_typology}{census_typology}

\item{minor_strata}{minor_strata}

\item{progress_fn}{progress_fn}
}
\description{
TBD
}
\author{
Johanna Herfaut
}
