% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_report.R
\name{compute_report}
\alias{compute_report}
\title{compute_report}
\usage{
compute_report(
  effort,
  effort_source = c("fisher_interview", "boat_counting", "household_interview"),
  landings,
  active_days = NULL,
  active_vessels = NULL,
  active_vessels_strategy = NULL,
  census_typology = NULL,
  minor_strata = NULL,
  validate = FALSE,
  progress_fn = NULL
)
}
\arguments{
\item{effort}{effort}

\item{effort_source}{effort_source}

\item{landings}{landings}

\item{active_days}{active_days}

\item{active_vessels}{active_vessels}

\item{active_vessels_strategy}{active_vessels_strategy}

\item{census_typology}{census_typology}

\item{minor_strata}{minor_strata. Default is \code{NULL}}

\item{validate}{validate}

\item{progress_fn}{a progress function with args (label, p). Default is \code{NULL}}
}
\value{
the result of Artfish
}
\description{
compute_report
}
