% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_statistics.R
\name{compute_summary_statistics}
\alias{compute_summary_statistics}
\title{Compute descriptive summary statistics}
\usage{
compute_summary_statistics(x, confidence = 0.9, keep_values = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{confidence}{Confidence level used to compute confidence intervals.Default is \code{0.90}.}

\item{keep_values}{Logical. Should the original values be returned in the output? Default is \code{TRUE}.}
}
\value{
A tibble containing descriptive statistics.
}
\description{
Computes a set of descriptive summary statistics from a numeric vector.
This function provides the generic engine used internally to compute descriptive statistics for specific indicators of interest
(e.g. \code{compute_cpue_per_trip_statistics()}, \code{compute_effort_activity_coefficient_per_observation_statistics()}).
}
