% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_utils.R
\name{get_artfish_results_for_ui}
\alias{get_artfish_results_for_ui}
\title{get_artfish_results_for_ui}
\usage{
get_artfish_results_for_ui(files, ref_fishing_units, ref_species)
}
\arguments{
\item{files}{list of computation output files}
}
\description{
get_artfish_results_for_ui
}
