% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{get_vrule_validators}
\alias{get_vrule_validators}
\title{Get data validators powered by \pkg{vrule}}
\usage{
get_vrule_validators()
}
\value{
the list of internal data validators as objects of class \link[vrule]{format_spec}
}
\description{
Get data validators powered by \pkg{vrule}
}
