% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{INFO}
\alias{INFO}
\title{util INFO logger}
\usage{
INFO(txt, ...)
}
\arguments{
\item{txt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{any values to be passed into \code{txt}. See \link{sprintf}}
}
\description{
INFO logger
}
