% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{util generic logger}
\usage{
logger(type = c("INFO", "WARN", "ERROR"), txt, ...)
}
\arguments{
\item{type}{type either INFO, WARN or DEBUG}

\item{txt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{any values to be passed into \code{txt}. See \link{sprintf}}
}
\description{
Generic logger
}
