% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny.R
\name{run_artfishr_shiny}
\alias{run_artfishr_shiny}
\alias{run_artfish_shiny}
\title{Runs a Artfishr Shiny application}
\usage{
run_artfish_shiny(...)
}
\arguments{
\item{...}{any optional args to pass to \link[shiny]{shinyApp}}
}
\description{
Runs a Artfishr Shiny application
}
