% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_plot.R
\name{sample_size_plot}
\alias{sample_size_plot}
\title{Produces a plot for the sample size}
\usage{
sample_size_plot(
  sample = NULL,
  population,
  sample.mean,
  sample.sd,
  alpha = 0.1
)
}
\arguments{
\item{sample}{sample}

\item{population}{population}

\item{sample.mean}{mean}

\item{sample.sd}{standard deviation}

\item{alpha}{pvalue of t distribution}
}
\description{
Produces a plot for the sample size
}
