% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{set_vrule_validators}
\alias{set_vrule_validators}
\title{Set data validators powered by \pkg{vrule}}
\usage{
set_vrule_validators()
}
\description{
Set data validators powered by \pkg{vrule}
}
