% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{validate_active_days}
\alias{validate_active_days}
\title{Validator for active days data}
\usage{
validate_active_days(data)
}
\arguments{
\item{data}{active days data}
}
\description{
Validator for active days data
}
