% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{validate_active_vessels}
\alias{validate_active_vessels}
\title{Validator for active vessels data}
\usage{
validate_active_vessels(data)
}
\arguments{
\item{data}{active vessels data}
}
\description{
Validator for active vessels data
}
