% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{validate_artfish_datasets}
\alias{validate_artfish_datasets}
\title{Wrapper functions for validating artfishr input data}
\description{
Family of functions to validate empty input dataset.
}
\details{
These functions are convenience wrappers around [validate_input_datasets()].
}
\seealso{
[validate_input_datasets()]
}
