% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{validate_effort}
\alias{validate_effort}
\title{Validator for effort data}
\usage{
validate_effort(data, effort_source = c("boat_counting", "fisher_interview"))
}
\arguments{
\item{data}{effort data}

\item{effort_source}{effort_source}
}
\description{
Validator for effort data
}
