% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_data_validators.R
\name{validate_landings}
\alias{validate_landings}
\title{Validator for landings data}
\usage{
validate_landings(data)
}
\arguments{
\item{data}{landings data}
}
\description{
Validator for landings data
}
