% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{WARN}
\alias{WARN}
\title{util WARN logger}
\usage{
WARN(txt, ...)
}
\arguments{
\item{txt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{any values to be passed into \code{txt}. See \link{sprintf}}
}
\description{
WARN logger
}
