% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_estimates_by_fleet_segment.R
\name{artfish_estimates_by_fleet_segment}
\alias{artfish_estimates_by_fleet_segment}
\title{artfish_estimates_by_fleet_segment}
\usage{
artfish_estimates_by_fleet_segment(
  data_effort = NULL,
  data_landing = NULL,
  ref_fishingunits = NULL,
  ref_species,
  year = NULL,
  month = NULL
)
}
\arguments{
\item{data_effort}{effort data}

\item{data_landing}{landing data}

\item{ref_fishingunits}{fishing units reference dataset}

\item{ref_species}{species reference dataset}
}
\value{
the result of Artfish
}
\description{
artfish_estimates_by_fleet_segment
}
