% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_fishing_unit.R
\name{artfish_shiny_fishing_unit_server}
\alias{artfish_shiny_fishing_unit_server}
\title{ARTFISH Fishing Unit - Shiny Server Module}
\usage{
artfish_shiny_fishing_unit_server(
  id,
  lang = NULL,
  estimate,
  effort_source,
  minor_strata = NULL,
  opts = list()
)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}

\item{lang}{Optional language parameter.
Can be either:
\itemize{
  \item a character string (static language), or
  \item a reactive returning a character string (dynamic language)
}
If \code{NULL}, the current global language is used.
Default is \code{NULL}}

\item{estimate}{A reactive data frame aggregating the output of \code{artfish_compute}, enriched with human-readable labels.}

\item{effort_source}{Reactive haracter string indicating the type of effort source.
Must be either \code{"fisher_interview"} or \code{"boat_counting"}.}

\item{minor_strata}{Reactive haracter string targeting a column name considered as minor strata.
Not activated}

\item{opts}{a named list of options. 
For now limited to:
- \code{refresh_ui} that gives the capacity to inject a refresh UI button (for dynamic computation)
- \code{values_ui} that allows to hide the UI related to value measurement}
}
\description{
Server-side logic for the ARTFISH Fishing Unit dashboard.

The *"Catch and effort - Detailed by fishing unit"* module provides a dedicated workspace for exploring key output metrics derived from \code{artfish_compute_report}, with a focus on one or multiple fishing units.
It enables users to analyse the evolution of fishing activity over time, with filtering options for fishing units and time range.

The module is composed of:
\itemize{
  \item A selector panel to adjust the time range and fishing units to display
  \item A set of KPIs presenting aggregated metrics for the selected period
  \item Three boxed plots powered by \code{fdishinyr::generic_chart} allowing users to analyse ranked top items for:
  \itemize{
    \item (1) Species composition (by quantity)
    \item (2) Fishing unit composition (by quantity)
    \item (3) Fishing unit composition (by value)
  }
  \item Six boxed plots powered by \code{fdishinyr::generic_chart} allowing users to analyse key Artfish metrics:
  \itemize{
    \item (1) Total catch estimation
    \item (2) Catch per unit effort (CPUE)
    \item (3) Total nominal effort
    \item (4) Activity coefficient
    \item (5) Number of active vessels
    \item (6) Total value estimation
  }
}
}
