% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_shiny_report.R
\name{artfish_shiny_report_server}
\alias{artfish_shiny_report_server}
\title{ARTFISH Report - Shiny Server Module}
\usage{
artfish_shiny_report_server(
  id,
  lang = NULL,
  estimate,
  effort_source,
  minor_strata
)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}

\item{lang}{Optional language parameter. Can be either:
\itemize{
  \item A character string (static language), or
  \item A reactive expression returning a character string (dynamic language)
}
If \code{NULL}, the current global language is used.
Default is \code{NULL}.}

\item{estimate}{A data frame aggregating the output of \code{artfish_compute}, enriched with human-readable labels.}

\item{effort_source}{Character string indicating the type of effort source.
Must be either \code{"fisher_interview"} or \code{"boat_counting"}.}

\item{minor_strata}{Character string targeting a column name considered as minor strata.}
}
\description{
Server-side logic for the ARTFISH Report dashboard.

The *"Detailed results of Artfish computation – by stratum"* module provides an ergonomic view of key intermediate results and calculation steps underlying the Artfish algorithm for a given temporal unit.

The initial view displays a selector inviting the user to choose a year. If no year is available, the user must first compute at least one indicator using \code{artfish_compute_report}.
Successive dropdown lists then allow the user to select a month and a fishing unit before submitting the selection. 
A complete report is displayed on the right-hand side.

The module is composed of:
\itemize{
  \item A selector panel to choose the report to display by selecting year, month, and fishing unit
  \item A button to validate the selection
  \item A set of KPIs presenting the report status and the type of effort source
  \item A gauge chart presenting the overall accuracy quality of the indicator
  \item An ergonomic report view including intermediate steps and final indicators for the estimation of:
  \itemize{
    \item (1) Effort
    \item (2) Landings
    \item (3) Estimated catch by species
  }
}
}
