% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_accuracy.R
\name{compute_accuracy}
\alias{compute_accuracy}
\title{Computes Accuracy}
\usage{
compute_accuracy(
  activity_coefficient,
  effort_estimate,
  cpue,
  sui,
  minor_strata = NULL
)
}
\arguments{
\item{activity_coefficient}{activity_coefficient}

\item{effort_estimate}{effort_estimate}

\item{cpue}{cpue}

\item{minor_strata}{minor_strata}
}
\value{
a \link[tibble]{tibble} object giving the different accuracy by strata
}
\description{
TODO
}
