% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_catch_estimate.R
\name{compute_catch_estimate}
\alias{compute_catch_estimate}
\title{Computes catch estimate}
\usage{
compute_catch_estimate(effort_estimate, landings, minor_strata = NULL)
}
\arguments{
\item{effort_estimate}{effort estimate computed with \link{compute_effort_estimate}}

\item{landings}{landings}

\item{minor_strata}{minor_strata. Default is \code{NULL}}
}
\value{
a \link{tibble} giving the estimated catch by strata
}
\description{
Computes catch estimate
}
