% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_catch_estimates_by_species.R
\name{compute_catch_estimates_by_species}
\alias{compute_catch_estimates_by_species}
\title{Computes catch estimates by species}
\usage{
compute_catch_estimates_by_species(landings, catch_estimate)
}
\arguments{
\item{landings}{landings}

\item{catch_estimate}{result of catch estimate computed with \link{compute_catch_estimate}}
}
\value{
a \link{tibble}
}
\description{
Computes catch estimates by species
}
