% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_cpue.R
\name{compute_cpue}
\alias{compute_cpue}
\title{Computes CPUE}
\usage{
compute_cpue(landings, minor_strata = NULL)
}
\arguments{
\item{landings}{landings}

\item{minor_strata}{minor_strata}
}
\value{
a \link{tibble} giving CPUE per strata
}
\description{
Computes CPUE
}
