% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_effort_activity_coefficient.R
\name{compute_effort_activity_coefficient}
\alias{compute_effort_activity_coefficient}
\title{Computes effort activity coefficient}
\usage{
compute_effort_activity_coefficient(
  effort,
  effort_source = c("fisher_interview", "boat_counting"),
  minor_strata = NULL
)
}
\arguments{
\item{effort}{effort data}

\item{effort_source}{effort}

\item{minor_strata}{minor_strata. Default is \code{NULL}}
}
\value{
a \link[tibble]{tibble} object giving activity coefficient by strata
}
\description{
The activity coefficient is computed from the effort. Depending on the source of
effort data, the source information used is different:
- In the case of 'fisher_interview' the coefficient of activity is computed as the ratio 
between the \code{effort_fishing_duration} and the \code{effort_fishing_reference_period}.
- In the case of 'boat_counting' the coefficient of activity is computed as the ratio 
between the \code{fleet_engagement_number} and the \code{fleet_engagement_max}

The computation is performed grouped by a strata compound at minimum by the \code{year},
\code{month} and \code{fishing_unit}. This strata can be extended by adding one or more
columns with the \code{minor_strata} argument.

Note: Additional check are performed to remove data with NAs, and ensure data consistency
}
