% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_effort_estimate.R
\name{compute_effort_estimate}
\alias{compute_effort_estimate}
\title{Computes nominal effort estimate}
\usage{
compute_effort_estimate(
  effort,
  effort_source = c("fisher_interview", "boat_counting", "household_interview"),
  landings,
  active_days = NULL,
  active_vessels = NULL,
  active_vessels_strategy = c("latest", "closest"),
  census_typology = NULL,
  minor_strata = NULL,
  progress_fn = NULL
)
}
\description{
Computes nominal effort estimate
}
