% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_effort_estimate.R
\name{compute_effort_estimate}
\alias{compute_effort_estimate}
\title{Computes effort estimate}
\usage{
compute_effort_estimate(
  active_vessels,
  active_vessels_strategy = c("latest", "closest"),
  effort,
  effort_source,
  active_days,
  minor_strata = NULL
)
}
\arguments{
\item{active_vessels}{active vessels}

\item{active_vessels_strategy}{The strategy to associate the active vessels to
the effort based on time. Active vessels period does not match necessarily the 
periods of data (effort, landings), and can be reported either by year or by year/month. 
This parameter let decide which methodology should be used to select the active vessels 
based on time. 
It can be either "latest" (taking the latest period), "closest" ie the closest active vessels
in time, after or before the data period. In case 2 periods before/after are equally closer,
the latest in time before the data period will be taken.}

\item{effort}{effort data}

\item{effort_source}{effort_source (register_interview / boat_counting)}

\item{active_days}{active_days}

\item{minor_strata}{minor_strata}
}
\description{
Computes effort estimate
}
