% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_new.R
\name{compute_effort_estimate}
\alias{compute_effort_estimate}
\title{Computes effort estimate}
\usage{
compute_effort_estimate(active_vessels, effort, active_days)
}
\arguments{
\item{active_vessels}{active vessels}

\item{effort}{effort data}

\item{active_days}{active_days}
}
\description{
Computes effort estimate
}
