% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/artfish_compute_effort_estimate_with_IRD_Pechart.R
\name{compute_effort_estimate_with_IRD_Pechart}
\alias{compute_effort_estimate_with_IRD_Pechart}
\title{Computes nominal effort estimate}
\usage{
compute_effort_estimate_with_IRD_Pechart(
  effort,
  effort_source = c("household_interview"),
  landings,
  active_days,
  census_typology,
  minor_stratum = NULL,
  progress_fn = NULL
)
}
\description{
TBD
}
\author{
Johanna Herfaut
}
