% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_compute_sui.R
\name{compute_sui}
\alias{compute_sui}
\title{Computes sui}
\usage{
compute_sui(effort, landings, minor_strata = NULL)
}
\arguments{
\item{effort}{effort}

\item{landings}{landings}

\item{minor_strata}{minor_strata. Default is \code{NULL}}
}
\value{
a \link[tibble]{tibble} object giving the different accuracy by strata
}
\description{
Compute the sufficient uniformity index for effort and catch corresponding to the uniformity of sampling over the sampled days.
}
