% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_template_creator.R
\name{create_artfish_templates}
\alias{create_artfish_templates}
\alias{create_active_vessels_template}
\alias{create_active_days_template}
\alias{create_landings_template}
\alias{create_effort_template}
\title{Wrapper functions for creating ArtFishR templates}
\usage{
create_active_vessels_template(include_meta = FALSE, save_as = NULL)

create_active_days_template(include_meta = FALSE, save_as = NULL)

create_landings_template(include_meta = FALSE, save_as = NULL)

create_effort_template(
  effort_source = c("boat_counting", "fisher_interview"),
  include_meta = FALSE,
  save_as = NULL
)
}
\description{
Family of functions to create empty templates for artfishr dataset types.
}
\details{
These functions are convenience wrappers around [create_artfish_template()].
}
\seealso{
[create_artfish_template()], [decode_artfish_schema()]
}
