% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_active_days.R
\name{generate_active_days}
\alias{generate_active_days}
\title{Generates a \link{tibble} of active days.}
\usage{
generate_active_days(active_vessels, effort, landings, minor_strata = NULL)
}
\arguments{
\item{active_vessels}{active vessels table}

\item{effort}{effort table}

\item{landings}{landings table}

\item{minor_strata}{minor strata}
}
\value{
an object of class \link{tibble} give active days
}
\description{
Function that generates a table of active days for all periods. The unique list
of fishing units are inherited from available tables (active_vessels, effort, landings). In
the same way, the list of minor strata values will be inherited based on the minor_strata
columns available in data.
}
