% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_active_days.R
\name{generate_active_days}
\alias{generate_active_days}
\title{Generates a \link{tibble} of active days}
\usage{
generate_active_days(year, month, fishing_units)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{fishing_units}{one or more fishing unit(s)}
}
\value{
an object of class \link{tibble} give active days
}
\description{
Generates a \link{tibble} of active days
}
