% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artfish_utils.R
\name{join_guess_by}
\alias{join_guess_by}
\title{join_guess_by}
\usage{
join_guess_by(x, y)
}
\description{
Gives the common column names between two tables. Util to be used
in \pkg{dplyr} \code{*_join} calls \code{by} argument to avoid explicit messages
triggered by \pkg{dplyr} when guessing the columns (in case \code{by} is not specified)
param x a \link{data.frame} or \link[tibble]{tibble}
param y a \link{data.frame} or \link[tibble]{tibble}
}
